packages/ocrbridge-tesseract/pyproject.toml
[project]
name = "ocrbridge-tesseract"
version = "2.0.1"
description = "Tesseract OCR engine for OCR Bridge"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "ocrbridge-core>=3.1.0",
    "pytesseract>=0.3.13",
    "Pillow>=10.0.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
build = ["python-semantic-release>=10.5.2"]
test = ["pytest~=9.0"]

[dependency-groups]
dev = [
    "pytest~=9.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "python-semantic-release>=9.0",
]
test = ["pytest~=9.0"]

# Entry point for engine discovery
[project.entry-points."ocrbridge.engines"]
tesseract = "ocrbridge.engines.tesseract:TesseractEngine"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocrbridge"]

[tool.ruff]
exclude = ["typings/*"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.10"
include = ["src"]
exclude = ["tests"]
extraPaths = ["typings"]
reportMissingTypeStubs = "none"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests requiring Tesseract binary",
    "slow: marks tests as slow running",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.semantic_release]
build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""
version_toml = ["pyproject.toml:project.version"]

[tool.uv.sources]
ocrbridge-core = { path = "../ocrbridge-core" }


